backend/requirements.txt
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==4.1.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
databases==0.9.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
inflection==0.5.1
motor==3.7.1
multidict==6.7.0
packaging==25.0
passlib==1.7.4
propcache==0.4.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.4
starlette==0.49.3
stripe==14.0.0
twilio==9.8.7
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
